Move content query out of app component and csb data and bap sam.gov data queries out of user dashboard component, and into utilities file
