+++ app/client/src/routes/helpdesk.tsx
@@ -6,13 +6,12 @@ import icon from "uswds/img/usa-icons-bg/search--white.svg";
 import icons from "uswds/img/sprite.svg";
 // ---
 import { serverUrl, messages } from "../config";
-import { getData, postData, useCsbData } from "../utilities";
+import { getData, postData, useContentData, useCsbData } from "../utilities";
 import { useHelpdeskAccess } from "components/app";
 import { Loading } from "components/loading";
 import { Message } from "components/message";
 import { MarkdownContent } from "components/markdownContent";
 import { TextWithTooltip } from "components/tooltip";
-import { useContentData } from "components/app";
 import { useDialogActions } from "contexts/dialog";
 
 type FormType = "application" | "payment-request" | "close-out";
