+++ app/client/src/routes/newApplicationForm.tsx
@@ -8,6 +8,7 @@ import { serverUrl, messages } from "../config";
 import {
   BapSamEntity,
   postData,
+  useContentData,
   useCsbData,
   useBapSamData,
   getUserInfo,
@@ -16,7 +17,6 @@ import { Loading } from "components/loading";
 import { Message } from "components/message";
 import { MarkdownContent } from "components/markdownContent";
 import { TextWithTooltip } from "components/tooltip";
-import { useContentData } from "components/app";
 
 type FormioSubmission = {
   [field: string]: unknown;
